Use router2 implementation instead of router
