+++ pkg/routing/rule_test.go
@@ -9,32 +9,55 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestAppRule(t *testing.T) {
+func TestConsumeRule(t *testing.T) {
 	keepAlive := 2 * time.Minute
 	pk, _ := cipher.GenerateKeyPair()
-	rule := AppRule(keepAlive, 1, 2, pk, 4, 3)
+
+	rule := ConsumeRule(keepAlive, 1, pk, 2, 3)
 
 	assert.Equal(t, keepAlive, rule.KeepAlive())
-	assert.Equal(t, RuleApp, rule.Type())
-	assert.Equal(t, RouteID(2), rule.RouteID())
-	assert.Equal(t, pk, rule.RemotePK())
-	assert.Equal(t, Port(3), rule.RemotePort())
-	assert.Equal(t, Port(4), rule.LocalPort())
+	assert.Equal(t, RuleConsume, rule.Type())
+	assert.Equal(t, RouteID(1), rule.KeyRouteID())
+	assert.Equal(t, pk, rule.RouteDescriptor().DstPK())
+	assert.Equal(t, Port(3), rule.RouteDescriptor().DstPort())
+	assert.Equal(t, Port(2), rule.RouteDescriptor().SrcPort())
 
-	rule.SetRouteID(3)
-	assert.Equal(t, RouteID(3), rule.RouteID())
+	rule.SetKeyRouteID(4)
+	assert.Equal(t, RouteID(4), rule.KeyRouteID())
 }
 
 func TestForwardRule(t *testing.T) {
 	trID := uuid.New()
 	keepAlive := 2 * time.Minute
-	rule := ForwardRule(keepAlive, 2, trID, 1)
+	pk, _ := cipher.GenerateKeyPair()
+
+	rule := ForwardRule(keepAlive, 1, 2, trID, pk, 3, 4)
 
 	assert.Equal(t, keepAlive, rule.KeepAlive())
 	assert.Equal(t, RuleForward, rule.Type())
-	assert.Equal(t, RouteID(2), rule.RouteID())
-	assert.Equal(t, trID, rule.TransportID())
+	assert.Equal(t, RouteID(1), rule.KeyRouteID())
+	assert.Equal(t, RouteID(2), rule.NextRouteID())
+	assert.Equal(t, trID, rule.NextTransportID())
+	assert.Equal(t, pk, rule.RouteDescriptor().DstPK())
+	assert.Equal(t, Port(4), rule.RouteDescriptor().DstPort())
+	assert.Equal(t, Port(3), rule.RouteDescriptor().SrcPort())
 
-	rule.SetRouteID(3)
-	assert.Equal(t, RouteID(3), rule.RouteID())
+	rule.SetKeyRouteID(5)
+	assert.Equal(t, RouteID(5), rule.KeyRouteID())
+}
+
+func TestIntermediaryForwardRule(t *testing.T) {
+	trID := uuid.New()
+	keepAlive := 2 * time.Minute
+
+	rule := IntermediaryForwardRule(keepAlive, 1, 2, trID)
+
+	assert.Equal(t, keepAlive, rule.KeepAlive())
+	assert.Equal(t, RuleIntermediaryForward, rule.Type())
+	assert.Equal(t, RouteID(1), rule.KeyRouteID())
+	assert.Equal(t, RouteID(2), rule.NextRouteID())
+	assert.Equal(t, trID, rule.NextTransportID())
+
+	rule.SetKeyRouteID(3)
+	assert.Equal(t, RouteID(3), rule.KeyRouteID())
 }
